Move WrapMode/Opacity to osu.Framework.Graphics.Textures
