+++ osu.Framework.Tests/Visual/Containers/TestSceneFrontToBack.cs
@@ -16,7 +16,6 @@ using osuTK.Graphics;
 using osuTK.Graphics.ES30;
 using osu.Framework.Graphics.Textures;
 using osu.Framework.Graphics.Primitives;
-using osu.Framework.Graphics.OpenGL.Textures;
 using osu.Framework.Graphics.Rendering;
 
 namespace osu.Framework.Tests.Visual.Containers
