+++ osu.Framework/Graphics/OpenGL/Textures/TextureGL.cs
@@ -197,34 +197,4 @@ namespace osu.Framework.Graphics.OpenGL.Textures
                 Opacity = Opacity.Mixed;
         }
     }
-
-    public enum WrapMode
-    {
-        /// <summary>
-        /// No wrapping. If the texture is part of an atlas, this may read outside the texture's bounds.
-        /// </summary>
-        None = 0,
-
-        /// <summary>
-        /// Clamps to the edge of the texture, repeating the edge to fill the remainder of the draw area.
-        /// </summary>
-        ClampToEdge = 1,
-
-        /// <summary>
-        /// Clamps to a transparent-black border around the texture, repeating the border to fill the remainder of the draw area.
-        /// </summary>
-        ClampToBorder = 2,
-
-        /// <summary>
-        /// Repeats the texture.
-        /// </summary>
-        Repeat = 3,
-    }
-
-    public enum Opacity
-    {
-        Opaque,
-        Mixed,
-        Transparent,
-    }
 }
