+++ osu.Framework/Graphics/Sprites/Sprite.cs
@@ -10,7 +10,6 @@ using osuTK;
 using osu.Framework.Graphics.Shaders;
 using osu.Framework.Allocation;
 using osu.Framework.Layout;
-using osu.Framework.Graphics.OpenGL.Textures;
 
 namespace osu.Framework.Graphics.Sprites
 {
