+++ osu.Framework/Graphics/Textures/Opacity.cs
@@ -1,0 +1,12 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+namespace osu.Framework.Graphics.Textures
+{
+    public enum Opacity
+    {
+        Opaque,
+        Mixed,
+        Transparent,
+    }
+}
